requirements.txt
streamlit
requests
python-dotenv
huggingface_hub